+++ src/commands/analyze.ts
@@ -25,6 +25,7 @@ import {Command, CommandOptions} from './command';
 
 export class AnalyzeCommand implements Command {
   name = 'analyze';
+  aliases = [];
 
   description = 'Writes analysis metadata in JSON format to standard out';