+++ src/commands/build.ts
@@ -31,6 +31,7 @@ const logger = logging.getLogger('cli.command.build');
 
 export class BuildCommand implements Command {
   name = 'build';
+  aliases = [];
 
   description = 'Builds an application-style project';