+++ CHANGELOG.md
@@ -53,6 +53,8 @@
 
 - `init`: Propagate `description` from `init` to application templates in `index.html` meta tag.
 
+- **New Command Aliases**: Commands now support aliases. `polymer install` has been aliased under `polymer i`.
+
 ## v0.18.1 [04-25-2017]
 
 - `init` small template fixes.